requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
av==16.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.2
fastapi==0.124.2
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
piper-tts==1.3.0
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
shellingham==1.5.4
silero-vad==6.2.0
sounddevice==0.5.3
soundfile==0.13.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
